test(DurableClock): move regression coverage
